extract getRemoteUserID
