+++ server/api/calendar.go
@@ -27,27 +27,6 @@ func (api *api) CreateCalendar(calendar *remote.Calendar) (*remote.Calendar, err
 	return client.CreateCalendar(calendar)
 }
 
-func (api *api) CreateEvent(event *remote.Event, mattermostUserIDs []string) (*remote.Event, error) {
-
-	// invite non-mapped Mattermost
-	for id := range mattermostUserIDs {
-		userID := mattermostUserIDs[id]
-		_, err := api.UserStore.LoadUser(userID)
-		if err != nil {
-			if err.Error() == "not found" {
-				err = api.Poster.DM(userID, "You have been invited to an MS office calendar event but have not linked your account.  Feel free to join us by connecting your www.office.com using `/msoffice connect`")
-			}
-		}
-	}
-
-	client, err := api.MakeClient()
-	if err != nil {
-		return nil, err
-	}
-
-	return client.CreateEvent(event)
-}
-
 func (api *api) DeleteCalendar(calendarID string) error {
 	client, err := api.MakeClient()
 	if err != nil {
@@ -75,29 +54,4 @@ func (api *api) GetUserCalendars(userID string) ([]*remote.Calendar, error) {
 	//DEBUG just use me as the user, for now
 	me, err := client.GetMe()
 	return client.GetUserCalendars(me.ID)
-}
-
-func (api *api) GetUserTimezone(mattermostUserID string) (string, error) {
-	client, err := api.MakeClient()
-	if err != nil {
-		return "", err
-	}
-
-	var remoteUserID string
-	if api.user != nil && api.user.MattermostUserID == mattermostUserID {
-		remoteUserID = api.user.Remote.ID
-	} else {
-		u, storeErr := api.UserStore.LoadUser(mattermostUserID)
-		if storeErr != nil {
-			return "", storeErr
-		}
-		remoteUserID = u.Remote.ID
-	}
-
-	settings, err := client.GetUserMailboxSettings(remoteUserID)
-	if err != nil {
-		return "", err
-	}
-
-	return settings.TimeZone, nil
 }
