+++ server/api/event.go
@@ -3,7 +3,31 @@
 
 package api
 
-import "github.com/pkg/errors"
+import (
+	"github.com/pkg/errors"
+
+	"github.com/mattermost/mattermost-plugin-mscalendar/server/remote"
+)
+
+func (api *api) CreateEvent(event *remote.Event, mattermostUserIDs []string) (*remote.Event, error) {
+	// invite non-mapped Mattermost
+	for id := range mattermostUserIDs {
+		userID := mattermostUserIDs[id]
+		_, err := api.UserStore.LoadUser(userID)
+		if err != nil {
+			if err.Error() == "not found" {
+				err = api.Poster.DM(userID, "You have been invited to an MS office calendar event but have not linked your account.  Feel free to join us by connecting your www.office.com using `/msoffice connect`")
+			}
+		}
+	}
+
+	client, err := api.MakeClient()
+	if err != nil {
+		return nil, err
+	}
+
+	return client.CreateEvent(event)
+}
 
 func (api *api) AcceptEvent(eventID string) error {
 	client, err := api.MakeClient()
